+++ src/lib.rs
@@ -6,3 +6,4 @@ pub mod routes;
 pub mod startup;
 pub mod telemetry;
 pub mod util;
+pub mod types;
