+++ src/types/mod.rs
@@ -1,0 +1,3 @@
+pub mod user;
+
+pub use user::*;
